tests/STT-test/requirements.txt
fastapi
uvicorn[standard]
python-multipart
requests
openai-whisper
